+++ modules/store/src/main/scala/docspell/store/queries/QItem.scala
@@ -1,8 +1,9 @@
 package docspell.store.queries
 
 import bitpeace.FileMeta
-import cats.implicits._
 import cats.effect.Sync
+import cats.data.OptionT
+import cats.implicits._
 import fs2.Stream
 import doobie._
 import doobie.implicits._
@@ -15,6 +16,44 @@ import org.log4s._
 
 object QItem {
   private[this] val logger = getLogger
+
+  def moveAttachmentBefore(
+      itemId: Ident,
+      source: Ident,
+      target: Ident
+  ): ConnectionIO[Int] = {
+
+    // rs < rt
+    def moveBack(rs: RAttachment, rt: RAttachment): ConnectionIO[Int] =
+      for {
+        n <- RAttachment.decPositions(itemId, rs.position, rt.position)
+        k <- RAttachment.updatePosition(rs.id, rt.position)
+      } yield n + k
+
+    // rs > rt
+    def moveForward(rs: RAttachment, rt: RAttachment): ConnectionIO[Int] =
+      for {
+        n <- RAttachment.incPositions(itemId, rt.position, rs.position)
+        k <- RAttachment.updatePosition(rs.id, rt.position)
+      } yield n + k
+
+    (for {
+      _ <- OptionT.liftF(
+        if (source == target)
+          Sync[ConnectionIO].raiseError(new Exception("Attachments are the same!"))
+        else ().pure[ConnectionIO]
+      )
+      rs <- OptionT(RAttachment.findById(source)).filter(_.itemId == itemId)
+      rt <- OptionT(RAttachment.findById(target)).filter(_.itemId == itemId)
+      n <- OptionT.liftF(
+        if (rs.position == rt.position || rs.position + 1 == rt.position)
+          0.pure[ConnectionIO]
+        else if (rs.position < rt.position) moveBack(rs, rt)
+        else moveForward(rs, rt)
+      )
+    } yield n).getOrElse(0)
+
+  }
 
   case class ItemData(
       item: RItem,